backend/pyproject.toml
[project]
name = "chatbot-backend"
version = "1.0.0"
description = "AI Chatbot Platform Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "alembic>=1.13.1",
    "psycopg[binary,pool]>=3.1.0",
    "pydantic-settings>=2.1.0",
    "pydantic[email]>=2.5.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[argon2]>=1.7.4",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
    "redis>=5.0.1",
    "celery[redis]>=5.3.6",
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "httpx>=0.26.0",
    "beautifulsoup4>=4.12.0",
    "PyPDF2>=3.0.1",
    "pdf2image>=1.17.0",
    "pillow>=10.0.0",
    "python-magic>=0.4.14",
    "pyyaml>=6.0.1",
    "asyncpg>=0.31.0",
    "psycopg2-binary>=2.9.11",
    "datapizza-ai>=0.0.9",
    "datapizza-ai-clients-openai-like>=0.0.9",
    "ollama>=0.6.1",
    "ddgs>=2.0.0",
    "googlesearch-python>=1.2.5",
    "chardet>=5.2.0",
    "msgpack-asgi>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
    "faker>=22.5.0",
    "factory-boy>=3.3.0",
]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
